v.1.2

+++ config/pushall.php
@@ -1,5 +1,5 @@
 <?php
 return [
-    'id' => env('PUSHALL_PRIVATE_ID'),
+    'id' => (int)env('PUSHALL_PRIVATE_ID'),
     'key' => env('PUSHALL_PRIVATE_KEY')
 ];
